+++ packages/std/src/math/uint64.rs
@@ -1,8 +1,9 @@
+use forward_ref::forward_ref_binop;
 use schemars::JsonSchema;
 use serde::{de, ser, Deserialize, Deserializer, Serialize};
 use std::convert::{TryFrom, TryInto};
 use std::fmt::{self};
-use std::ops;
+use std::ops::{self, Rem};
 
 use crate::errors::{DivideByZeroError, OverflowError, OverflowOperation, StdError};
 use crate::Uint128;
@@ -224,6 +225,19 @@ impl<'a> ops::Div<&'a Uint64> for Uint64 {
         Self(self.u64().checked_div(rhs.u64()).unwrap())
     }
 }
+
+impl Rem for Uint64 {
+    type Output = Self;
+
+    /// # Panics
+    ///
+    /// This operation will panic if `rhs` is zero.
+    #[inline]
+    fn rem(self, rhs: Self) -> Self {
+        Self(self.0.rem(rhs.0))
+    }
+}
+forward_ref_binop!(impl Rem, rem for Uint64, Uint64);
 
 impl ops::Shr<u32> for Uint64 {
     type Output = Self;
@@ -612,5 +626,31 @@ mod tests {
             Uint64(u64::MAX - 1)
         );
         assert_eq!(Uint64(u64::MAX).wrapping_pow(2), Uint64(1));
+    }
+
+    #[test]
+    #[allow(clippy::op_ref)]
+    fn uint64_implements_rem() {
+        let a = Uint64::new(10);
+        assert_eq!(a % Uint64::new(10), Uint64::zero());
+        assert_eq!(a % Uint64::new(2), Uint64::zero());
+        assert_eq!(a % Uint64::new(1), Uint64::zero());
+        assert_eq!(a % Uint64::new(3), Uint64::new(1));
+        assert_eq!(a % Uint64::new(4), Uint64::new(2));
+
+        // works for refs
+        let a = Uint64::new(10);
+        let b = Uint64::new(3);
+        let expected = Uint64::new(1);
+        assert_eq!(a % b, expected);
+        assert_eq!(a % &b, expected);
+        assert_eq!(&a % b, expected);
+        assert_eq!(&a % &b, expected);
+    }
+
+    #[test]
+    #[should_panic(expected = "divisor of zero")]
+    fn uint64_rem_panics_for_zero() {
+        let _ = Uint64::new(10) % Uint64::zero();
     }
 }